+++ benchmarks/types/array/Head.js
@@ -4,12 +4,9 @@ require('../../../src/ts-tooling');
 
 (new Benchmark.Suite).add('ts-tooling Tail', () => {
     [1,2,3,4,5,6,7,8,9,10].Head(5);
-}).add('native loop', () => {
+}).add('native slice', () => {
     const tmp = [1,2,3,4,5,6,7,8,9,10];
-    const res = [];
-    for (let i = 0; i < tmp.length; i++) {
-        res.push(tmp[i]);
-    }
+    tmp.slice(0, 5);
 }).add('lodash takeRight', () => {
     _.take([1,2,3,4,5,6,7,8,9,10], 5);
 }).on('cycle', function(event) {